+++ sdk/keyvault/azkeys/CHANGELOG.md
@@ -10,8 +10,8 @@
 ### Breaking Changes
 * Changed the `Data` to `EncodedPolicy` on the `KeyReleasePolicy` struct.
 * Changed the `Updated`, `Created`, and `Expires` properties to `UpdatedOn`, `CreatedOn`, and `ExpiresOn`.
-* Renamed `JSONWebKeyOperation` to `KeyOperation`.
-* Renamed `JSONWebKeyCurveName` to `KeyCurveName`
+* Renamed `JSONWebKeyOperation` to `Operation`.
+* Renamed `JSONWebKeyCurveName` to `CurveName`
 * Prefixed all KeyType constants with `KeyType`
 * Changed `KeyBundle` to `KeyVaultKey` and `DeletedKeyBundle` to `DeletedKey`
 * Renamed `KeyAttributes` to `KeyProperties`